sentiment/requirements.txt
finnhub-python
tweepy
scipy
matplotlib
numpy
google-cloud-datastore
google-cloud-language